+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 3.x
 ===========
 - Remove Deprecated ContainerAwareCommand (by @kejwmen)
 - Update travis matrix for current symfony versions
+- Remove deprecations from phpunit
 
 Version 3.0
 ===========